렌트카 엔티티 작업완료
