+++ howThereProject/src/main/java/com/howthere/app/embed/Address.java
@@ -1,0 +1,11 @@
+package com.howthere.app.embed;
+
+import javax.persistence.Embeddable;
+
+@Embeddable
+public class Address {
+
+    private String address;
+    private Double Latitude;
+    private Double Longitude;
+}
