--- howThereProject/src/main/java/com/howthere/app/entity/Address.java
@@ -1,11 +1,0 @@
-package com.howthere.app.entity;
-
-import javax.persistence.Embeddable;
-
-@Embeddable
-public class Address {
-
-    private String address;
-    private Double rentCarLatitude;
-    private Double rentCarLongitude;
-}
