+++ howThereProject/src/main/java/com/howthere/app/entity/file/RentCarFile.java
@@ -1,14 +1,11 @@
 package com.howthere.app.entity.file;
 
-import com.howthere.app.entity.RentCar;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
 
 import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 
 @Entity
