--- howThereProject/src/main/java/com/howthere/app/entity/rentCarType/RentCarType.java
@@ -1,5 +1,0 @@
-package com.howthere.app.entity.rentCarType;
-
-public enum  RentCarType {
-    경차, 소형차, 중형차, 대형차, SUV
-}
