+++ howThereProject/src/main/java/com/howthere/app/type/RentCarType.java
@@ -1,0 +1,5 @@
+package com.howthere.app.type;
+
+public enum  RentCarType {
+    LiteCar, CompactCar , mediumSizedCar ,  LargeCar, SUV
+}
